Corrections carried over from C++11 implementation to C++14 and C++17

+++ BoundedTypes/CoreFunctions/Library/SafeTypes_CXX14.hpp
@@ -41,13 +41,13 @@ namespace RomanoViolet
   class SafeType< NumeratorForMinBound, 1, NumeratorForMaxBound, 1 >
   {
   public:
-    SafeType( int value );
-    int getValue( );
+    SafeType( float value );
+    float getValue( );
 
   private:
     int _min;
     int _max;
-    int _value;
+    float _value;
 
     // define temporary data structure to hold new numerators and denominators if these need to be
     // transformed.
